Add CocoaAsyncSocket

+++ HalliGalli/Classes/Model/RoomInfo.swift
@@ -9,14 +9,15 @@
 import UIKit
 
 class RoomInfo: Codable {
+    
     /// 房间ID
-    var roomID: Int64?
+    var roomID: Int?
     /// 房间是否开始
     var isstarted:Bool
     /// 当前人数
-    var count: Int64?
+    var count: Int?
     
-    init(isstarted:Bool,roomID: Int64, count: Int64) {
+    init(isstarted:Bool,roomID: Int, count: Int) {
         self.isstarted = isstarted
         self.roomID = roomID
         self.count = count
